Keras Check
Erin trained the model in a later version of Keras than what I had which caused errors when I was running my code. So, this model was trained with a lower version of Keras. If you get a "KeyError: 0" error running the code with the later version, then, move the files in this folder to the same level as the app.py file and see if that works.