HTML B4 Flask
This is the HTML before the flask was created. Interestingly, the scribble pad changed when the drawingboard,js library was moved over to the static folder.  This is just a reference for the original look and feel of the scribble pad. The Predict Button is NOT attached to a flask.